refactor: erase bg func
