+++ src/views/DesignEditor/components/Toolbox/LayerPanel/BgLayer/BgLayer.tsx
@@ -7,6 +7,7 @@ import ColorPicker from "~/components/UI/ColorPicker/ColorPicker"
 import { useEditor, useFrame } from "@layerhub-io/react"
 import { backgroundLayerType, checkboxBGUrl, deviceUploadType } from "~/constants/contants"
 import UploadImgModal from "~/components/UI/UploadImgModal/UploadImgModal"
+import { EraseBgFunc } from "~/views/DesignEditor/utils/functions/EraseBgFunc"
 
 const BgLayer = ({ showLayer, handleClose }: any) => {
   const [activeState, setActiveState] = useState(-1)
@@ -52,54 +53,6 @@ const BgLayer = ({ showLayer, handleClose }: any) => {
     } else setActiveState(idx)
   }
 
-  const eraseBg = () => {
-    const bgObject = editor?.frame?.background?.canvas?._objects.filter(
-      (el: any) => el.metadata?.type === backgroundLayerType
-    )[0]
-
-    editor?.frame.resize({ width: frame.width, height: frame.height })
-
-    const deviceUploadImg = editor?.frame?.background?.canvas?._objects.filter(
-      (el: any) => el.metadata?.type === deviceUploadType
-    )[0]
-
-    const backgroundImage = editor?.frame?.background?.canvas?._objects.filter(
-      (el: any) => el?.type === "BackgroundImage"
-    )[0]
-
-    editor?.frame?.setBackgroundColor("#FFF")
-
-    if (!bgObject && !deviceUploadImg) {
-      const options = {
-        type: "BackgroundImage",
-        src: checkboxBGUrl,
-        preview: checkboxBGUrl,
-        metadata: { generationDate: new Date().getTime(), type: backgroundLayerType },
-      }
-
-      editor.objects.add(options).then(() => {
-        setTimeout(() => {
-          editor.objects.setAsBackgroundImage()
-        }, 100)
-      })
-    } else if (deviceUploadImg || backgroundImage) {
-      if (deviceUploadImg) editor.objects.removeById(deviceUploadImg.id)
-      else if (backgroundImage) editor.objects.removeById(backgroundImage.id)
-      const options = {
-        type: "BackgroundImage",
-        src: checkboxBGUrl,
-        preview: checkboxBGUrl,
-        metadata: { generationDate: new Date().getTime(), type: backgroundLayerType },
-      }
-
-      editor.objects.add(options).then(() => {
-        setTimeout(() => {
-          editor.objects.setAsBackgroundImage()
-        }, 100)
-      })
-    }
-  }
-
   const colors = ["#FF6BB2", "#B69DFF", "#30C5E5", "#7BB872", "#49A8EE", "#3F91A2", "#DA4F7A", "#FFFFFF"]
 
   return showLayer ? (
@@ -129,9 +82,7 @@ const BgLayer = ({ showLayer, handleClose }: any) => {
                 classes.box,
                 " pointer d-flex justify-content-center align-items-center flex-column ml-1 pointer"
               )}
-              onClick={() => {
-                eraseBg()
-              }}
+              onClick={() => EraseBgFunc({ editor })}
             >
               <Icons.TrashIcon size={"20"} />
               <p>Delete</p>
